move xmlname::validate_and_extract to xpath::eval
